requirements.txt
Flask
Gunicorn
Pillow
blinker==1.8.2
click==8.1.7
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
Werkzeug==3.0.3